@ppedrot 2022-09-29
How do we know we will not fall in the same pitfall as #15944, namely a compiler that does not see that the casts can be eliminated? In that PR according to @silene it seemed to be version-specific.