@ppedrot 2022-09-29
> it doesn't deserve to call itself an optimizing compiler.

The same argument holds for Obj.magic, and nonetheless they were not eliminated.