@silene 2022-09-29
From a quick test with various compiler versions (as far back as 4.04), it seems like OCaml alwaysoptimizes this code construct properly.