@ejgallego 2022-09-29
> it doesn't deserve to call itself an optimizing compiler.

I'm not sure the OCaml compiler without flambda calls itself an optimizing compiler.