@SkySkimmer 2022-09-30
>This patch doesn't exactly look like a simplification to my eyes. Can you elaborate on the motivation in the PR description?

rectypes is bad for development, several times I had partial code and merlin found some nonsense recursive type instead of producing a helpful error.