@ppedrot 2022-09-30
I also support the removal of rectypes, regardless of the other issues with native compute.