@Alizter 2022-09-30
+1 for removing rectypes