@silene 2022-09-30
> Can you elaborate on the motivation in the PR description?

Removing `-rectypes` is also important because the OCaml compiler considers it to be viral. As a consequence, if your program depends on a Coq module, you need to compile it with `-rectypes` too. This is an endless source of pain. For example, we are (were?) recommending that users link their extracted programs against some Coq modules (e.g., for native int63 support), which was just impossible because of `-rectypes`.