@maximedenes 2022-09-30
> Removing `-rectypes` is also important because the OCaml compiler considers it to be viral. As a consequence, if your program depends on a Coq module, you need to compile it with `-rectypes` too. This is an endless source of pain. For example, we are (were?) recommending that users link their extracted programs against some Coq modules (e.g., for native int63 support), which was just impossible because of `-rectypes`.

Right, I hit this kind of issues in the past, but had forgotten about it.

So indeed, if we manage to get rid of it without impacting too much the readability of the compiler code, or the performance of the generated code, it sounds like a nice improvement.