@maximedenes 2022-09-30
@SkySkimmer You could try to increment `n` in `RecNoAlloc`, to get a more meaningful runtime.