@SkySkimmer 2022-09-30
Maybe I measured wrong, when I run `Eval native_compute in foo`, undo then redo is the second run the runtime or just getting the result from cache?
With n=30 I get 7.6s on the first run and unchanged after (both master and this PR).