@maximedenes 2022-09-30
If `foo` is a global definition, you should get it from the cache the second time (after undo+redo).