@SkySkimmer 2022-10-01
fourcolor has some slowdowns (typically 0.5s, 1.5%) on the qed of things that look like
~~~coq
Lemma red189to206 : reducible_in_range 189 206 the_configs.
Proof.
 CheckReducible.
Qed.
~~~
AFAICT CheckReducible https://github.com/coq-community/fourcolor/blob/944732e8b47b77e704b344315a14eec9cbfd9c26/theories/cfreducible.v#L96-L98 uses VM but not native, so I don't know what's going on.

rewriter has slowdowns (0.6s, 2% to 5%) on stuff looking like
~~~coq
Time Definition args_of_size_by_sample (k : kind_of_rewrite) (s : size)
  := Eval native_compute in eta_size (s' => eta_kind (k' => args_of_size_by_sample' k' s') k) s.
~~~