@ppedrot 2022-10-05
@SkySkimmer what about the fiat-crypto failure? It does look like it's due to native compilation failing, but with the new workers one cannot be sure about anything.