@SkySkimmer 2022-10-05
That's an old runner, but it failed elsewhere eg https://gitlab.com/coq/coq/-/jobs/3118572051